Rename: auth 관련 타입 정리
