+++ src/types/Auth/AuthDataRequest.ts
@@ -1,0 +1,6 @@
+export interface AuthDataRequest {
+  name?: string;
+  email: string;
+  password: string;
+  passwordChk?: string;
+}
