+++ src/types/Auth/AuthDataResponse.ts
@@ -1,0 +1,5 @@
+export interface AuthDataResponse {
+  statusCode: number;
+  data: object;
+  timestamp: string;
+}
