+++ src/types/Auth/AuthType.ts
@@ -1,0 +1,1 @@
+export type AuthType = '로그인' | '회원가입';
